@@ -16,19 +16,21 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.0.0-beta.5",
-    "@angular/animations": "^6.1.0",
+    "@angular/animations": "8.0.0",
     "@angular/cdk": "~7.3.3",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/core": "^6.1.0",
+    "@angular/common": "8.0.0",
+    "@angular/compiler": "8.0.0",
+    "@angular/core": "8.0.0",
     "@angular/fire": "^5.2.1",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
+    "@angular/forms": "8.0.0",
+    "@angular/http": "8.0.0-beta.10",
     "@angular/material": "^7.3.3",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
+    "@angular/platform-browser": "8.0.0",
+    "@angular/platform-browser-dynamic": "8.0.0",
+    "@angular/router": "8.0.0",
     "@auth0/angular-jwt": "^2.0.0",
+    "@nstudio/nativescript-floatingactionbutton": "^1.0.0",
+    "@nstudio/nativescript-pulltorefresh": "^1.0.1",
     "@types/jquery": "^3.3.29",
     "amazon-cognito-auth-js": "^1.2.4",
     "amazon-cognito-js": "^1.1.0",
@@ -42,38 +44,36 @@
     "firebase": "^6.3.1",
     "hammerjs": "^2.0.8",
     "material-design-icons": "^3.0.1",
-    "nativescript-angular": "~7.2.0",
+    "nativescript-angular": "8.0.2",
     "nativescript-background-http": "^3.4.1",
     "nativescript-braintree": "^2.0.4",
     "nativescript-dialog": "^0.1.0",
     "nativescript-dom": "^2.0.2",
     "nativescript-facebook": "^4.1.0",
-    "nativescript-floatingactionbutton": "^5.1.0",
     "nativescript-google-maps-sdk": "^2.7.0",
     "nativescript-imagecropper": "^1.0.6",
     "nativescript-imagepicker": "^6.2.0",
     "nativescript-masked-text-field": "^4.0.2",
     "nativescript-modal-datetimepicker": "^1.2.0",
     "nativescript-plugin-firebase": "^9.0.4",
-    "nativescript-pulltorefresh": "^2.3.0",
     "nativescript-ripple": "^2.2.1",
     "nativescript-theme-core": "~1.0.4",
-    "nativescript-ui-dataform": "^4.0.1",
-    "nativescript-ui-sidedrawer": "^6.0.0",
+    "nativescript-ui-dataform": "5.0.0",
+    "nativescript-ui-sidedrawer": "7.0.0",
     "ngx-braintree": "^6.1.4",
     "ngx-paypal": "^5.0.0-beta.10",
     "ngx-stripe": "^7.2.0",
     "reflect-metadata": "~0.1.12",
-    "rxjs": "~6.2.0",
+    "rxjs": "6.3.3",
     "socket.io": "^2.2.0",
     "timers": "^0.1.1",
-    "tns-core-modules": "~5.2.0",
-    "zone.js": "~0.8.26"
+    "tns-core-modules": "6.0.1",
+    "zone.js": "0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.8.0",
+    "@angular-devkit/build-angular": "0.800.3",
     "@angular/cli": "~6.2.4",
-    "@angular/compiler-cli": "^6.1.0",
+    "@angular/compiler-cli": "8.0.0",
     "@angular/language-service": "^6.1.0",
     "@nativescript/schematics": "^0.5.2",
     "@types/jasmine": "^2.8.16",
@@ -87,16 +87,16 @@
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "nativescript-dev-webpack": "^0.20.0",
+    "nativescript-dev-webpack": "1.0.1",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~2.9.2"
+    "typescript": "3.4.5"
   },
   "nativescript": {
     "id": "org.nativescript.collegepool",
     "tns-android": {
-      "version": "5.3.1"
+      "version": "6.0.0"
     }
   }
 }
