@@ -16,8 +16,7 @@ public class Continuation extends java.lang.Object
   public java.lang.Object then(com.google.android.gms.tasks.Task<java.lang.Object> param_0) {
     java.lang.Object[] args = new java.lang.Object[1];
     args[0] = param_0;
-    return (java.lang.Object)
-        com.tns.Runtime.callJSMethod(this, "then", java.lang.Object.class, args);
+    return com.tns.Runtime.callJSMethod(this, "then", java.lang.Object.class, args);
   }
 
   public int hashCode__super() {
